refact(backend): Consolidate Design to use feature
Instead of have a Design Functional view, we change to a Desing Feature View.
That means which Controllers, Models, Services and etc are grouped by Features.
