+++ backend/src/BlogDoFt.SbusEmulatorViewer.Api/Features/ServiceBus/Models/EntitiesParser.cs
@@ -1,28 +1,24 @@
-using Microsoft.Extensions.FileProviders;
+using BlogDoFt.SbusEmulatorViewer.Api.Models;
 using Microsoft.Extensions.Options;
 using System.Text.Json;
 
-namespace BlogDoFt.SbusEmulatorViewer.Api.Models;
+namespace BlogDoFt.SbusEmulatorViewer.Api.Features.ServiceBus.Models;
 
 public class EntitiesParser
 {
     private readonly string _configPath;
-    private readonly IFileProvider _fileProvider;
 
     private readonly Lazy<Task<IReadOnlyList<Topic>>> _lazyTopics;
 
     public EntitiesParser(
-        IOptions<ServiceBusSettings> sbusSettings,
-        IWebHostEnvironment env)
+        IOptions<ServiceBusSettings> sbusSettings)
     {
         _configPath = sbusSettings.Value.EmulatorConfigFile;
-        _fileProvider = env.ContentRootFileProvider;
         _lazyTopics = new Lazy<Task<IReadOnlyList<Topic>>>(LoadTopicsAsync);
     }
 
     public Task<IReadOnlyList<Topic>> GetTopicsAsync(CancellationToken cancellationToken = default)
         => _lazyTopics.Value;
-
 
     private async Task<IReadOnlyList<Topic>> LoadTopicsAsync()
     {
@@ -36,7 +32,7 @@ public class EntitiesParser
         var root = await JsonSerializer
             .DeserializeAsync<RootConfig>(stream, new JsonSerializerOptions
             {
-                PropertyNameCaseInsensitive = true
+                PropertyNameCaseInsensitive = true,
             });
 
         var namespaces = root?.UserConfig?.Namespaces ?? [];
@@ -46,41 +42,30 @@ public class EntitiesParser
             .SelectMany(ns => ns.Topics!)
             .Select(t => new Topic(
                 TopicName: t.Name!,
-                Subscriptions: t.Subscriptions?
-                    .Select(s => s.Name!)
-                    .ToArray()
-                ?? []
-            ))
+                Subscriptions: t.Subscriptions?.Select(s => s.Name!).ToArray() ?? []))
             .ToList()
             .AsReadOnly();
 
         return topics;
     }
 
-
-    private sealed class SubscriptionConfig
+    private sealed record SubscriptionConfig(string? Name)
     {
-        public string? Name { get; set; }
     }
 
-    private sealed class TopicConfig
+    private sealed record TopicConfig(string? Name, SubscriptionConfig[]? Subscriptions)
     {
-        public string? Name { get; set; }
-        public SubscriptionConfig[]? Subscriptions { get; set; }
     }
 
-    private sealed class NamespaceConfig
+    private sealed record NamespaceConfig(TopicConfig[]? Topics)
     {
-        public string? Name { get; set; }
-        public TopicConfig[]? Topics { get; set; }
-    }
-    private sealed class UserConfig
-    {
-        public NamespaceConfig[]? Namespaces { get; set; }
     }
 
-    private sealed class RootConfig
+    private sealed record UserConfig(NamespaceConfig[]? Namespaces)
     {
-        public UserConfig? UserConfig { get; set; }
+    }
+
+    private sealed record RootConfig(UserConfig? UserConfig)
+    {
     }
 }
