+++ backend/src/BlogDoFt.SbusEmulatorViewer.Api/Features/ServiceBus/Models/MessageDetails.cs
@@ -1,0 +1,8 @@
+namespace BlogDoFt.SbusEmulatorViewer.Api.Features.ServiceBus.Models;
+
+public class MessageDetails
+{
+    public object? Body { get; set; }
+
+    public object? Details { get; set; }
+}
